Merged Reactor & latests fixes
Updated a few functions to static
Removed dead code (casts, imports, etc.)
Fixed overlapped properties with local variables/parameters
Fixed parameters modifications
Added a few null-check
Added missing @Override

Updated WarpDrive block description to WarpDriveCore (players were
confused)
Updated cloaking core (removing frequency, add enable state)
Fixed Peripheral.equals checks
Fixed Camera falling
Added dirty fix for IC2 support in new reactor
Disabled oxygen for creative players and dead/invulnerable entities
Added tentative oxygen level display, increased drown damage

+++ src/cr0s/WarpDrive/EntityStarCore.java
@@ -38,11 +38,11 @@ public final class EntityStarCore extends Entity
     {
         super(world);
         this.xCoord = x;
-        this.posX = (double) x;
+        this.posX = x;
         this.yCoord = y;
-        this.posY = (double) y;
+        this.posY = y;
         this.zCoord = z;
-        this.posZ = (double) z;
+        this.posZ = z;
         this.radius = radius;
     }
